automation_scripts/requirements-agent.txt
azure-identity>=1.19.0
databricks-sdk>=0.40.0